+++ Source/TheBoxSoftware.Reflection/Syntax/CSharp/CSharpFormatter.cs
@@ -137,7 +137,7 @@ namespace TheBoxSoftware.Reflection.Syntax.CSharp
                 {
                     tokens.AddRange(FormatTypeDetails(details.ArrayOf));
                     tokens.Add(Constants.ArrayStart);
-                    tokens.Add(new SyntaxToken(new String(',', details.ArrayShape.Rank - 1), SyntaxTokens.Text));
+                    tokens.Add(new SyntaxToken(new String(',', (int)details.ArrayShape.Rank - 1), SyntaxTokens.Text));
                     tokens.Add(Constants.ArrayEnd);
                 }
             }
